Refactor Ui class out
The Ui class prints out messages to
users in the command line.

However,the project has changed from a command line one to
Gui, this function is no longer necessary.

Get response by Duke using getResponse function in Duke

Do it this way so that we can convert response by Duke to Labels in
order to display to user in Gui

Let's delete all occurrences of Ui class in all methods to make code
deater
