+++ src/main/java/duke/main/Main.java
@@ -2,6 +2,7 @@ package duke.main;
 
 import java.io.IOException;
 
+import duke.ui.MainWindow;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
