+++ src/main/resources/view/DialogBox.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.HBox?>
 <fx:root xmlns:fx="http://javafx.com/fxml/1" alignment="TOP_RIGHT" maxWidth="1.7976931348623157E308" prefWidth="400.0"
-         type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/18">
+         type="javafx.scene.layout.HBox" xmlns="http://javafx.com/javafx/8.0.171">
     <children>
         <Label fx:id="dialog" graphicTextGap="10.0" maxHeight="1.7976931348623157E308" prefWidth="273.0"
                style="-fx-label-padding: 10; -fx-border-radius: 10;" text="Label" wrapText="true"
